requirements.txt
functions-framework==3.*
google-cloud-storage
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.94.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0
flask==2.3.2
pyarrow

